测试

+++ util.sh
@@ -235,8 +235,11 @@ mv -f /tmp/${zipname} ${deployDir}/ || exit 1
 cd ${deployDir} || exit 1
 unzip -o ${zipname} || exit 1
 
-mv -f ${deployDir}/${CD_ZIP_ROOT}/*.zip ${deployDir}/ || exit 1
-rm -rf ${deployDir}/${CD_ZIP_ROOT}/ || exit 1
+if [[ ${CD_ZIP_ROOT} != "" ]]; then
+    mv -f ${deployDir}/${CD_ZIP_ROOT}/*.zip ${deployDir}/ || exit 1
+    rm -rf ${deployDir}/${CD_ZIP_ROOT}/ || exit 1
+fi
+
 rm -f ${zipname} || exit 1
 
 echo "#遍历目录"
